Fix search suggestions.
The search suggestions offered in examples drop down for check status and notification history are invalid; the drop-down suggests searches such as `"check" == "my check"` or `"notification rule" == "my rule"`, but, neither of those searches will ever return any results, because `"check"` and "notification rule"` are not column names used in the underlying queries.

This PR proposes altering the suggestions to match the existing queries. Arguably, it might be a better user experience to alert the underlying queries to match the existing suggestions (`check` is nicer to type than `checkName`), but this was a simpler change.